requirements.txt
fastapi==0.115.2
uvicorn==0.32.0
pydantic==2.9.2
python-dotenv==1.0.1
rapidfuzz==3.14.3
httpx==0.28.1
anthropic==0.86.0
groq==1.1.1
elevenlabs==2.39.0
